Ajuste nomenclatura, criação do cord config e criação do service para transferencia.

+++ ms-benefit-module/src/main/java/com/example/backend/domain/model/BenefitDTO.java
@@ -5,6 +5,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @AllArgsConstructor
 @Builder
 @Data
@@ -13,7 +15,7 @@ public class BenefitDTO {
     private Long id;
     private String name;
     private String description;
-    private Long value;
+    private BigDecimal value;
     private boolean active;
     private Long version;
 }
